requirements.txt
asgiref
certifi
charset-normalizer
coreapi
coreschema
Django
django-autoslug
django-rest-swagger
djangorestframework
idna
itypes
Jinja2
MarkupSafe
openapi-codec
pillow
python-dotenv
requests
simplejson
sqlparse
tzdata
uritemplate
urllib3
